#37576 Fix "Trying to access array offset on value of type null" - tree already deleted.

+++ components/ILIAS/Tree/classes/class.ilMaterializedPathTree.php
@@ -316,11 +316,15 @@ class ilMaterializedPathTree implements ilTreeImplementation
                 $a_node_id,
                 $this->getTree()->getTreeId()
             ));
-            $row = $this->db->fetchAssoc($res);
+            $node = $this->db->fetchAssoc($res);
+
+            if($node === null) {
+                return; //Nothing to delete. $node does not exists
+            }
 
             $query = 'DELETE FROM ' . $this->getTree()->getTreeTable() . ' ' .
-                'WHERE path BETWEEN ' . $this->db->quote($row['path'], 'text') . ' ' .
-                'AND ' . $this->db->quote($row['path'] . '.Z', 'text') . ' ' .
+                'WHERE path BETWEEN ' . $this->db->quote($node['path'], 'text') . ' ' .
+                'AND ' . $this->db->quote($node['path'] . '.Z', 'text') . ' ' .
                 'AND ' . $this->getTree()->getTreePk() . ' = ' . $this->db->quote(
                     $this->getTree()->getTreeId(),
                     'integer'
